+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+
+### Added
+
+ - defined input and output schema
+
 ### Changed
 
 - raise OSError on post result graph error
+- removed write_md and produce_graph bool parameters
 
 ## [1.0.0beta4] 2024-07-12